+++ CHANGELOG.md
@@ -2,13 +2,12 @@
 
 All notable changes to the Wazuh app project will be documented in this file.
 
-## Wazuh v4.12.0 - OpenSearch Dashboards 2.16.0 - Revision 00
+## Wazuh v4.12.0 - OpenSearch Dashboards 2.19.0 - Revision 00
 
 ### Added
 
 - Support for Wazuh 4.12.0
 - Add setting to limit the number of rows in csv reports [#7182](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7182)
-- Added a test to check the tables columns fields are known and new fields are added to the known fields of alerts index [#7268](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7268)
 
 ### Changed
 
@@ -21,8 +20,19 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Fixed width of long fields in the document detail flyout [#7206](https://github.com/wazuh/wazuh-dashboard-plugins/issues/7206)
 - Fixed logging UI logs due to logger undefined property [#7267](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7267)
 - Fixed TOP-5-SO filter management in Endpoints > Summary [#7278](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7278)
+- Fixed CSV export not filtering by timerange [#7304](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7304)
 
-## Wazuh v4.11.0 - OpenSearch Dashboards 2.16.0 - Revision 01
+## Wazuh v4.11.1 - OpenSearch Dashboards 2.16.0 - Revision 00
+
+### Added
+
+- Support for Wazuh 4.11.1
+
+### Fixed
+
+- Fixed documentation links related to agent management [#7299](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7299)
+
+## Wazuh v4.11.0 - OpenSearch Dashboards 2.16.0 - Revision 03
 
 ### Added